requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.2
openpyxl>=3.1.2
python-dateutil>=2.9.0
python-dotenv>=1.0.1
requests>=2.32.3
